+++ CHANGELOG.md
@@ -1,11 +1,13 @@
 
-## [5.2.10](https://github.com/jinghui70/rainbow-dbaccess/compare/v5.2.9...v5.2.10) (2024-10-21)
+## [5.2.11](https://github.com/jinghui70/rainbow-dbaccess/compare/v5.2.9...v5.2.11) (2024-11-16)
 
 ### Features
 
-- 增加 where(boolean condition, Cnds cnds) 这类函数
+- 增加 where(boolean condition, () -> Cnds) 这类函数
 - 增加 ClobObjectFieldMapper
-- TreeUtil 遍历时先遍历 children
+- 增加 Bool 到 YN 或 TF 的映射
+- ObjectSql Update 的 set 函数，用 FieldMapper 转换值
+- TreeUtil 的遍历函数，增加是否先序遍历的参数
 
 ## [5.2.9](https://github.com/jinghui70/rainbow-dbaccess/compare/v5.2.8...v5.2.9) (2024-10-21)